rasp-requirements.txt
absl-py==2.2.2
Adafruit-Blinka==8.56.0
adafruit-circuitpython-busdevice==5.2.11
adafruit-circuitpython-connectionmanager==3.1.3
adafruit-circuitpython-framebuf==1.6.7
adafruit-circuitpython-requests==4.1.10
adafruit-circuitpython-ssd1305==1.4.0
adafruit-circuitpython-ssd1306==2.12.19
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.77.0
Adafruit-PureIO==1.1.11
astunparse==1.6.3
attrs==25.3.0
binho-host-adapter==0.1.6
cachetools==5.5.2
cbor2==5.6.5
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
contourpy==1.3.2
cycler==0.12.1
flatbuffers==20181003210633
fonttools==4.57.0
gast==0.4.0
google-auth==2.39.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.71.0
h5py==3.13.0
idna==3.10
jax==0.4.30
jaxlib==0.4.30
keras==2.12.0
kiwisolver==1.4.8
lgpio==0.2.2.0
libclang==18.1.1
luma.core==2.4.2
luma.oled==3.14.0
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.1
mediapipe==0.10.15
ml_dtypes==0.5.1
numpy==1.23.5
oauthlib==3.2.2
opencv-contrib-python==4.11.0.86
opt_einsum==3.4.0
packaging==25.0
pillow==11.2.1
protobuf==4.25.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyftdi==0.56.0
pyparsing==3.2.3
pyserial==3.5
python-dateutil==2.9.0.post0
pyusb==1.3.1
requests==2.32.3
requests-oauthlib==2.0.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.1
rsa==4.9.1
scipy==1.15.2
six==1.17.0
smbus2==0.5.0
sounddevice==0.5.1
spidev==3.6
sysv-ipc==1.1.0
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow==2.12.0
tensorflow-cpu-aws==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
typing_extensions==4.13.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.14.1